% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediationPlot.R
\name{add_ellipse}
\alias{add_ellipse}
\title{Make an ellipse}
\usage{
add_ellipse(x = 0, y = 0, theta = 0, height = height, width = width,
  npoints = 100, color = "black", fill = "white", ...)
}
\arguments{
\item{x}{An integer indicating x position}

\item{y}{An integer indicating y position}

\item{theta}{An integer}

\item{height}{An integer}

\item{width}{An integer}

\item{npoints}{An integer}

\item{color}{A string indicating color}

\item{fill}{A string indicating color}

\item{...}{further arguments to be passed to geom_polygon}
}
\description{
Make an ellipse
}

