% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEquation.R
\name{addCovarEquation}
\alias{addCovarEquation}
\title{Add covariates to equation}
\usage{
addCovarEquation(equation, covar = list(), prefix = "h")
}
\arguments{
\item{equation}{The equation}

\item{covar}{A list}

\item{prefix}{prefix}
}
\description{
Add covariates to equation
}
