% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semPlot1.R
\name{addlabel}
\alias{addlabel}
\title{Add a label to the semMediation plot}
\usage{
addlabel(x1 = 0, y1 = 0, x2 = 10, y2 = 10, height = 3, width = 5,
  start = "ellipse", end = "rect", label = label, useLabel, ...)
}
\arguments{
\item{x1}{The x coordinate of start point}

\item{y1}{The y coordinate of start point}

\item{x2}{The x coordinate of end point}

\item{y2}{The y coordinate of end point}

\item{height}{A number indicating height of the rectangle}

\item{width}{A number indicating width of the rectangle}

\item{start}{The start geom. Either "rect" or "ellipse"}

\item{end}{The end geom. Either "rect" or "ellipse"}

\item{label}{The label passed on to geom_label()}

\item{useLabel}{Whether use geom_label instead of geom_label. Default value is FALSE.}

\item{...}{Other arguments passed on to geom_label()}
}
\description{
Add a label to the semMediation plot
}

