% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semPlot1.R
\name{addline}
\alias{addline}
\title{Add a line to the semMediation plot}
\usage{
addline(x1 = 0, y1 = 0, x2 = 10, y2 = 10, height = 3, width = 5,
  start = "ellipse", end = "rect", linetype = "solid", size = 0.2,
  colour = "black", ...)
}
\arguments{
\item{x1}{The x coordinate of start point}

\item{y1}{The y coordinate of start point}

\item{x2}{The x coordinate of end point}

\item{y2}{The y coordinate of end point}

\item{height}{A number indicating height of the rectangle}

\item{width}{A number indicating width of the rectangle}

\item{start}{The start geom. Either "rect" or "ellipse"}

\item{end}{The end geom. Either "rect" or "ellipse"}

\item{linetype}{The linetype passed to geom_segment()}

\item{size}{the size of geom_segment()}

\item{colour}{the colour of geom_segment()}

\item{...}{Other arguments passed on to geom_segment()}
}
\description{
Add a line to the semMediation plot
}

