% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediationPlot.R
\name{addLinetype}
\alias{addLinetype}
\title{Add line type}
\usage{
addLinetype(res)
}
\arguments{
\item{res}{A data.frame. Result of parameterEstimates function of package lavaan or subset.}
}
\description{
Add line type
}

