% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADHD.R
\docType{data}
\name{ADHD}
\alias{ADHD}
\title{Data Set for Teacher's Intervention for ADHD}
\format{A data.frame with 334 rows and 12 variables:
\describe{
  \item{id}{study identification number}
  \item{gender}{male=1, female=2}
  \item{age}{age groups 1-4}
  \item{general}{general knowledge estimate derived from Knowledge of
  Attention Deficit Disorder Scale(KADDS) developed by Sciutto, Terjesen and Frank(2000). Maximum 15 points.}
  \item{symptoms}{knowledge about ADHD symptoms estimate derived from Knowledge
  of Attention Deficit Disorder Scale(KADDS) developed by Sciutto, Terjesen and Frank(2000). Maximum 9 points.}
  \item{treatmt}{knowledge about ADHD treatment estimate derived from Knowledge of Attention Deficit Disorder Scale(KADDS)
  developed by Sciutto, Terjesen and Frank(2000). Maximum 12 points.}
  \item{cognitiv}{estimates of cognitive empathy. Maximum 35 points.}
  \item{emotion}{estimates of emotional empathy. Maximum 35 points.}
  \item{disposit}{estimates of dispositional empathy. Maximum 60 points.}
  \item{attitude}{estimates of attitude empathy. Maximum 35 points.}
  \item{classrm}{estimates of classroom intervention. Maximum 36 points.}
  \item{instruct}{estimates of instructional intervention. Maximum 36 points.}
}}
\source{
{Effects of teacher's knowledge and empathy on educational intervention for ADHD: Focused on the mediating effet of empathy. J Korean Acad Psychiatr Ment Health Nurs 2013:22;45-55.}
}
\usage{
ADHD
}
\description{
A dataset contains measures about the teacher's knowlege, empathy and intervention about attention-deficit hyperactivity disorder(ADHD).
}
\keyword{datasets}

