% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediationPlot.R
\name{adjustPos}
\alias{adjustPos}
\title{Adjust position for arrows}
\usage{
adjustPos(df, maxx = 80, maxy = 30, height = 3, width = 5)
}
\arguments{
\item{df}{A data.frame}

\item{maxx}{maximum x position}

\item{maxy}{maximum y position}

\item{height}{A number indicating height of the rectangle}

\item{width}{A number indicating width of the rectangle}
}
\description{
Adjust position for arrows
}

