% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticalDiagram.R
\name{changeLabelName}
\alias{changeLabelName}
\title{Change Label Names}
\usage{
changeLabelName(x, labels)
}
\arguments{
\item{x}{A character vector}

\item{labels}{A list}
}
\description{
Change Label Names
}
