% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptDiagram.R
\name{conceptDiagram}
\alias{conceptDiagram}
\title{Make concept Diagram}
\usage{
conceptDiagram(X = "X", M = "M", Y = "Y", latent = rep(FALSE, 3),
  xb = FALSE, radx = 0.12, rady = 0.05, xmargin = 0.03,
  yinterval = NULL, moderator = list(), labels = list())
}
\arguments{
\item{X}{character Name of independent variable}

\item{M}{character Name of mediator variable}

\item{Y}{character Name of dependent variable}

\item{latent}{Logical. whether or not X,Y and Z are latent variables or not}

\item{xb}{Logical. if positive draw linew between X and (Y+Z)}

\item{radx}{horizontal radius of the box.}

\item{rady}{vertical radius of the box.}

\item{xmargin}{horizontal margin of plot}

\item{yinterval}{vertical interval bewteen box}

\item{moderator}{optional lists of moderator}

\item{labels}{optionallabels of X,Y and Z variables}
}
\description{
Make concept Diagram
}
\examples{
labels=list(X="Time Spent in\\n Grad School", M="# of\\n Publications", Y="# of Job Offers")
conceptDiagram(xb=TRUE,labels=labels)
moderator=list(name="Z1",label="Time Spent\\n with Alex",pos=3,
    site=list(c("a","b","c")),latent=FALSE)
conceptDiagram(moderator=moderator,labels=labels)
moderator=list(name=c("Z1","Z2"),label=c("Time Spent\\n with Alex","Z2label"),pos=c(3,3),
    site=list(c("a","b","c"),c("b","c")),latent=c(FALSE,FALSE))
conceptDiagram(moderator=moderator,labels=labels,yinterval=0.4)
}
