% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condEffect.R
\name{condEffect}
\alias{condEffect}
\title{Make interaction plot}
\usage{
condEffect(data = mtcars, Y = "mpg", X = "wt", M = "am",
  probs = c(0.16, 0.5, 0.84), showEffect = TRUE)
}
\arguments{
\item{data}{A data.frame}

\item{Y}{The name of dependent variable}

\item{X}{The name of independent variable}

\item{M}{The name of modifier variable}

\item{probs}{A vector of probability weights for obtaining the elements of the vector being sampled.
Default value is c(0.16,0.5,0.84)}

\item{showEffect}{A logical}
}
\description{
Make interaction plot
}
\examples{
condEffect(data=mtcars,Y="mpg",X="wt",M="am")
condEffect(data=mtcars,Y="mpg",X="wt",M="hp")
}
