% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorPlot.R
\name{CorPlot}
\alias{CorPlot}
\title{Draw correlation plot}
\usage{
CorPlot(fit, label = 2, yreverse = TRUE, ...)
}
\arguments{
\item{fit}{An object of class lavaan. Result of sem function of package lavaan}

\item{label}{if 0, no label(default), if 1, use r value as label, if 2, use r value with significant mark as label}

\item{yreverse}{Logical. if true, reverse the order of y axis.}

\item{...}{Further arugement to be passed on to ggCor}
}
\value{
A ggplot
}
\description{
Draw correlation plot
}
\examples{

# require(mycor)
# require(ggplot)
# require(ggiraphExtra)
# require(lavaan)

# model='
# knowledge =~ general+symptoms+treatmt
# empathy =~ cognitiv+emotion+disposit+attitude
# intervention =~ classrm+instruct
# '
# mediationModel=makeEquation(X="knowledge",M="empathy",Y="intervention")
# model=paste0(model,mediationModel)
# fit=sem(model,data=ADHD)
# CorPlot(fit)
}

