% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorPlot.R
\name{corTable}
\alias{corTable}
\title{Make a table with correlation}
\usage{
corTable(fit, vanilla = TRUE, addFooter = FALSE)
}
\arguments{
\item{fit}{An object of class lavaan. Result of sem function of package lavaan}

\item{vanilla}{Logical. If true, vanilla.table is returned}

\item{addFooter}{Logical. If true, footer added}
}
\description{
Make a table with correlation
}

