% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediationPlot.R
\name{distributePos}
\alias{distributePos}
\title{Make a positions}
\usage{
distributePos(start, end, count)
}
\arguments{
\item{start}{A number indicating the start position}

\item{end}{A number indicating the end position}

\item{count}{A number indicating the count of items}
}
\description{
Make a positions
}

