% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediationPlot.R
\name{ellipse}
\alias{ellipse}
\title{Make a data.frame fof ellipse}
\usage{
ellipse(hlaxa = 1, hlaxb = 1, theta = 0, xc = 0, yc = 0,
  npoints = 100)
}
\arguments{
\item{hlaxa}{An integer}

\item{hlaxb}{An integer}

\item{theta}{An integer}

\item{xc}{An integer indicating center of x position}

\item{yc}{An integer indicating center of y position}

\item{npoints}{An integer}
}
\description{
Make a data.frame fof ellipse
}

