% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorPlot.R
\name{estimatesTable2}
\alias{estimatesTable2}
\title{convert parameterEstimates to flextable}
\usage{
estimatesTable2(fit, vanilla = FALSE, ...)
}
\arguments{
\item{fit}{An object of class lavaan. Result of sem function of package lavaan}

\item{vanilla}{Logical}

\item{...}{Further argumant to be passed to estimatesTable()}
}
\description{
convert parameterEstimates to flextable
}
