% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semDiagram.R
\name{extractNumber}
\alias{extractNumber}
\title{Extract Numbers in character vectors}
\usage{
extractNumber(x)
}
\arguments{
\item{x}{a numeric or character vector}
}
\description{
Extract Numbers in character vectors
}
