% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticalDiagram.R
\name{findNames}
\alias{findNames}
\title{convert a vextor of names with list}
\usage{
findNames(labels, names)
}
\arguments{
\item{labels}{A named list}

\item{names}{A character vector to look for}
}
\description{
convert a vextor of names with list
}
