% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediationPlot.R
\name{fit2df}
\alias{fit2df}
\title{Make a data.frame for mediationPlot}
\usage{
fit2df(fit)
}
\arguments{
\item{fit}{An object of class lavaan. Result of sem function of package lavaan}
}
\description{
Make a data.frame for mediationPlot
}

