% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semPlot1.R
\name{get_startx}
\alias{get_startx}
\title{Get start x  position}
\usage{
get_startx(x1, y1, x2, y2, height, width, start = "rect")
}
\arguments{
\item{x1}{The x coordinate of start point}

\item{y1}{The y coordinate of start point}

\item{x2}{The x coordinate of end point}

\item{y2}{The y coordinate of end point}

\item{height}{A number indicating height of the rectangle}

\item{width}{A number indicating width of the rectangle}

\item{start}{The start geom. Either "rect" or "ellipse"}
}
\description{
Get start x  position
}

