% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modmedEquation.R
\name{interactStr}
\alias{interactStr}
\title{make interaction equation}
\usage{
interactStr(x, prefix = "a")
}
\arguments{
\item{x}{character vector}

\item{prefix}{prefix}
}
\description{
make interaction equation
}
