% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediationPlot.R
\name{jitterPos}
\alias{jitterPos}
\title{Make a positions}
\usage{
jitterPos(center, step, count)
}
\arguments{
\item{center}{A number indicating the center position}

\item{step}{A number indicating the intervals betweeon items}

\item{count}{A number indicating the count of items}
}
\description{
Make a positions
}

