% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semDiagram.R
\name{makeDiagram}
\alias{makeDiagram}
\title{Make diagram euation}
\usage{
makeDiagram(fit, whatLabels = "std", mediationOnly = FALSE,
  regression = TRUE, indirect = FALSE, secondIndirect = FALSE,
  total = FALSE, residuals = FALSE, nodesep = NULL, ranksep = NULL)
}
\arguments{
\item{fit}{A data.frame. Result of parameterEstimates function of package lavaan}

\item{whatLabels}{What should the edge labels indicate in the path diagram? Choices are c("est","std","name").}

\item{mediationOnly}{Whether or not draw mediation effect only. Default value is FALSE.}

\item{regression}{Whether or not draw regression. Default value is TRUE.}

\item{indirect}{Whether or not draw indirect effects. Default value is FALSE.}

\item{secondIndirect}{Whether or not draw 2nd indirect effects. Default value is FALSE.}

\item{total}{Whether or not draw total effect. Default value is FALSE.}

\item{residuals}{Logical, should residuals (and variances) be included in the path diagram?}

\item{nodesep}{separation between nodes, in inches}

\item{ranksep}{separation between ranks, in inches.}
}
\description{
Make diagram euation
}
