% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripleInteraction.R
\name{makeIndirectEquation}
\alias{makeIndirectEquation}
\title{Make indirect equation}
\usage{
makeIndirectEquation(X, M, temp1, temp2, temp3, moderatorNames,
  range = FALSE)
}
\arguments{
\item{X}{A character string}

\item{M}{A character string}

\item{temp1}{A character vector}

\item{temp2}{A character vector}

\item{temp3}{A character vector}

\item{moderatorNames}{A character vector}

\item{range}{A logical}
}
\description{
Make indirect equation
}
\examples{
X="negemot";M="ideology"
temp1=c("a1*negemot","a2*sex","a4*negemot*sex","a5*negemot*age","a6*sex*age")
temp2="b1*ideology"
temp3="c1*negemot"
moderatorNames=c("age","sex")
cat(makeIndirectEquation(X,M,temp1,temp2,temp3,moderatorNames))
cat(makeIndirectEquation(X,M,temp1,temp2,temp3,moderatorNames,range=TRUE))
}
