% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semDiagram.R
\name{makeSameRank}
\alias{makeSameRank}
\title{make same rank}
\usage{
makeSameRank(equation, temp, addOrder = TRUE)
}
\arguments{
\item{equation}{semequation to add}

\item{temp}{character vector}

\item{addOrder}{logical. Whether or not add order}
}
\description{
make same rank
}
