% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediationPlot.R
\name{mediationPlot}
\alias{mediationPlot}
\title{Make a data.frame for mediationPlot}
\usage{
mediationPlot(fit, maxx = 60, maxy = 30, height = 5, width = 5,
  whatLabels = "std", useLabel = FALSE, usecolor = TRUE, clean = TRUE,
  base_size = 5, base_family = "Arial", mediationOnly = FALSE,
  residuals = TRUE, regression = TRUE, indirect = FALSE,
  secondIndirect = FALSE, total = FALSE, mode = 1)
}
\arguments{
\item{fit}{A data.frame. Result of parameterEstimates function of package lavaan}

\item{maxx}{An integer indicating maximum x position}

\item{maxy}{An integer indicating maximum y position}

\item{height}{A number indicating height of the rectangle}

\item{width}{A number indicating width of the rectangle}

\item{whatLabels}{What should the edge labels indicate in the path diagram? Choices are c("est","std","name").}

\item{useLabel}{Whether use geom_label instead of geom_text. Default value is FALSE.}

\item{usecolor}{Whether use colors for variables. Default value is TRUE.}

\item{clean}{Whether use theme_clean. Default value is TRUE.}

\item{base_size}{An integer indicating the font size.}

\item{base_family}{A character indicating base font family}

\item{mediationOnly}{Whether or not draw mediation effect only. Default value is FALSE.}

\item{residuals}{Whether or not draw residuals(and variance). Default value is TRUE.}

\item{regression}{Whether or not draw regression. Default value is TRUE.}

\item{indirect}{Whether or not draw indirect effects. Default value is FALSE.}

\item{secondIndirect}{Whether or not draw 2nd indirect effects. Default value is FALSE.}

\item{total}{Whether or not draw total effect. Default value is FALSE.}

\item{mode}{plot mode. 1 or 2.}
}
\description{
Make a data.frame for mediationPlot
}

