% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediationPlot.R
\name{mergeDataPos}
\alias{mergeDataPos}
\title{Merge Data with position}
\usage{
mergeDataPos(res, df2, whatLabels = "est", width = 5, height = 3)
}
\arguments{
\item{res}{A data.frame. Result of parameterEstimates function of package lavaan or subset.}

\item{df2}{A data.grame}

\item{whatLabels}{What should the edge labels indicate in the path diagram? Choices are c("est","std","name").}

\item{width}{A number indicating width of the rectangle}

\item{height}{A number indicating height of the rectangle}
}
\description{
Merge Data with position
}

