% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediationPlot.R
\name{mergeDataPos2Ind}
\alias{mergeDataPos2Ind}
\title{Merge Data with position for the 2nd indirect effect}
\usage{
mergeDataPos2Ind(res2Ind, res, df2, whatLabels = "est", width = 5,
  height = 3)
}
\arguments{
\item{res2Ind}{A data.frame.A subset of result of parameterEstimates function of package lavaan}

\item{res}{A data.frame. Result of parameterEstimates function of package lavaan or subset.}

\item{df2}{A data.grame}

\item{whatLabels}{What should the edge labels indicate in the path diagram? Choices are c("est","std","name").}

\item{width}{A number indicating width of the rectangle}

\item{height}{A number indicating height of the rectangle}
}
\description{
Merge Data with position for the 2nd indirect effect
}

