% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsSummaryTable.R, R/modelsSummaryTable2.R
\name{modelsSummary}
\alias{modelsSummary}
\title{Make Model Coef Summary}
\usage{
modelsSummary(fit1, fit2)

modelsSummary(fit1, fit2)
}
\arguments{
\item{fit1}{object of class lm}

\item{fit2}{object of class lm}

\item{fit1}{object of class lm}

\item{fit2}{object of class lm}
}
\value{
A data.frame

A data.frame
}
\description{
Make Model Coef Summary

Make Model Coef Summary
}
