% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsSummary.R
\name{modelsSummaryTable}
\alias{modelsSummaryTable}
\title{Make Model Coef Table}
\usage{
modelsSummaryTable(x)
}
\arguments{
\item{x}{object of class modelSummary}
}
\value{
A flextable
}
\description{
Make Model Coef Table
}
\examples{
fit1=lm(mpg~wt,data=mtcars)
fit2=lm(mpg~wt*hp,data=mtcars)
fit3=lm(mpg~wt*hp*am,data=mtcars)
x=modelsSummary(list(fit1))
modelsSummaryTable(x)
x=modelsSummary(list(fit1,fit2))
modelsSummaryTable(x)
x=modelsSummary(list(fit1,fit2,fit3))
modelsSummaryTable(x)
}
