% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modmedEquation.R
\name{modmedEquation}
\alias{modmedEquation}
\title{Make moderated mediation equation}
\usage{
modmedEquation(X = "", M = "", Y = "", moderator = list(),
  labels = NULL, range = FALSE)
}
\arguments{
\item{X}{A character vectors indicating independent variables}

\item{M}{A character vectors indicating mediators}

\item{Y}{A character vectors indicating dependent variables}

\item{moderator}{moderator}

\item{labels}{labels}

\item{range}{Whether or not add range equation}
}
\description{
Make moderated mediation equation
}
\examples{
X="X";M="M";Y="Y"
moderator=list(name=c("w","z"),site=list(c("a"),c("a")))
cat(modmedEquation(X=X,M=M,Y=Y,moderator=moderator))
}
