% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptDiagram.R
\name{myarrow}
\alias{myarrow}
\title{Draw arrow}
\usage{
myarrow(from, to, lwd = 1, adjust = 1, label = "", arr.pos = NULL,
  ...)
}
\arguments{
\item{from}{coordinates (x,y) of the point *from* which to draw arrow.}

\item{to}{coordinates (x,y) of the point *to* which to draw arrow.}

\item{lwd}{line width}

\item{adjust}{adjust position}

\item{label}{label}

\item{arr.pos}{arrow position}

\item{...}{Further argument to be passed to straightarrow()}
}
\description{
Draw arrow
}
