% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsSummaryTable.R, R/modelsSummaryTable2.R
\name{pformat}
\alias{pformat}
\title{Make p value format}
\usage{
pformat(x)

pformat(x)
}
\arguments{
\item{x}{A numeric vector}

\item{x}{A numeric vector}
}
\description{
Make p value format

Make p value format
}
