% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADHD.R, R/education.R
\docType{data}
\name{pmacro}
\alias{pmacro}
\title{Data Set for process macro
A dataset for support for process macro}
\format{A data.frame with 6 variables
\describe{
\item{no}{A number of process macro model}
\item{X}{Name of the independent variable}
\item{M}{Name of the mediator}
\item{Y}{Name of the independent variable}
\item{modName}{Name(s) of moderator names collapsed with colon}
\item{modSite}{Site(s) of moderation collapsed with colon}
\item{pos}{Position of moderator(s) in a diagram collapsed with comma}
}}
\usage{
pmacro

pmacro
}
\description{
Data Set for process macro
A dataset for support for process macro

Data Set for process macro model
}
\keyword{datasets}
