% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmacroModel.R
\name{pmacroModel}
\alias{pmacroModel}
\title{draw conceptual diagram of process macro model}
\usage{
pmacroModel(no = 1, labels = list(), covar = list())
}
\arguments{
\item{no}{process macro model number}

\item{labels}{A character list}

\item{covar}{A optional list of covariates}
}
\description{
draw conceptual diagram of process macro model
}
\examples{
covar=list(name=c("C1","C2","C3"),label=c("ese","sex","tenure"),site=list("M",c("M","Y"),c("Y")))
pmacroModel(4,covar=covar)
pmacroModel(1,covar=covar)
pmacroModel(1)
}
