% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliabilityTable.R
\name{reliabilityTable}
\alias{reliabilityTable}
\title{make reliability Table}
\usage{
reliabilityTable(fit, vanilla = FALSE)
}
\arguments{
\item{fit}{An object of a class lavaan}

\item{vanilla}{Logical}
}
\description{
make reliability Table
}
