% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condEffect.R
\name{relpos}
\alias{relpos}
\title{get relative position og a ggplot object}
\usage{
relpos(p, pos = c(0.5, 0.9))
}
\arguments{
\item{p}{A ggplot object}

\item{pos}{A numeric vector of xposition and y position}
}
\description{
get relative position og a ggplot object
}
