% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripleInteraction.R
\name{seekNameVars}
\alias{seekNameVars}
\title{select names of variables from list var}
\usage{
seekNameVars(vars, site = "a")
}
\arguments{
\item{vars}{A list}

\item{site}{Site for look kor}
}
\description{
select names of variables from list var
}
\examples{
vars=list(name=list(c("sex","age"),c("ese","sex")),site=list(c("a","b"),"c"))
seekNameVars(vars,"c")
}
