% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semDiagram.R
\name{selectEven}
\alias{selectEven}
\title{Select characters contains even numbers}
\usage{
selectEven(x)
}
\arguments{
\item{x}{a numeric or character vector}
}
\description{
Select characters contains even numbers
}
