% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semDiagram.R
\name{selectOdd}
\alias{selectOdd}
\title{Select characters contains odd numbers}
\usage{
selectOdd(x)
}
\arguments{
\item{x}{a numeric or character vector}
}
\description{
Select characters contains odd numbers
}
