% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semDiagram.R
\name{semDiagram}
\alias{semDiagram}
\title{Make sem diagram}
\usage{
semDiagram(fit, ...)
}
\arguments{
\item{fit}{A data.frame. Result of parameterEstimates function of package lavaan}

\item{...}{Further arguments tobe passed to makeDiagram()}
}
\description{
Make sem diagram
}
