% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediationPlot.R
\name{theme_clean}
\alias{theme_clean}
\title{Make a clean theme for ggplot}
\usage{
theme_clean(base_size = 12, base_family = "NanumGothic")
}
\arguments{
\item{base_size}{An integer indicating base font size}

\item{base_family}{A character indicating base font family}
}
\description{
Make a clean theme for ggplot
}

