% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripleInteraction.R
\name{treatModerator}
\alias{treatModerator}
\title{Treat moderator name with mean value}
\usage{
treatModerator(ind, moderatorNames)
}
\arguments{
\item{ind}{An equation}

\item{moderatorNames}{character vectors}
}
\description{
Treat moderator name with mean value
}
\examples{
ind="(a1+a4*sex+a5*age)*(b1)"
moderatorNames=c("age","sex")
treatModerator(ind,moderatorNames)
}
