% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripleInteraction.R
\name{tripleEquation}
\alias{tripleEquation}
\title{Make equation with triple interaction}
\usage{
tripleEquation(M = NULL, Y = NULL, vars, dep, suffix = 0, covar,
  mode = 0)
}
\arguments{
\item{M}{Name of mediator}

\item{Y}{Name of dependent variable}

\item{vars}{variable names to be interact}

\item{dep}{name of dependent variable}

\item{suffix}{A number}

\item{covar}{A list of covariates}

\item{mode}{A number}
}
\description{
Make equation with triple interaction
}
\examples{
vars=c("negemot","sex","age");dep="govact";suffix=0
covar=list(name=c("C1","C2","C3"),label=c("ese","sex","tenure"),site=list(c("M","Y"),"Y","Y"))
Y="govact"
tripleEquation(Y=Y,vars=vars,dep=dep,covar=covar)
tripleEquation(Y=Y,vars=vars,dep=dep,covar=covar,mode=1)
}
