% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripleInteraction.R
\name{tripleInteraction}
\alias{tripleInteraction}
\title{Make triple interaction equation}
\usage{
tripleInteraction(vars, dep, prefix = "c", suffix = 0, mode = 0)
}
\arguments{
\item{vars}{variable names to be interact}

\item{dep}{name of dependent variable}

\item{prefix}{A character}

\item{suffix}{A number}

\item{mode}{A number}
}
\description{
Make triple interaction equation
}
\examples{
vars=c("negemot","sex","age");dep="govact"
tripleInteraction(vars,dep)
tripleInteraction(vars,dep,mode=1)
}
