% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorPlot.R
\name{df2Flextable}
\alias{df2Flextable}
\title{convert data.frame to Flextable}
\usage{
df2Flextable(df, bg = "#5B7778")
}
\arguments{
\item{df}{a data.frame}

\item{bg}{background color of header}
}
\value{
A Flextable
}
\description{
convert data.frame to Flextable
}

