% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsSummaryTable.R
\name{getInfo}
\alias{getInfo}
\title{Get information of a model}
\usage{
getInfo(fit, digits = 3)
}
\arguments{
\item{digits}{integer indicating the number of decimal places}

\item{fit1}{object of class lm}
}
\description{
Get information of a model
}
