% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semDiagram.R
\name{makeDiagram}
\alias{makeDiagram}
\title{Make diagram euation}
\usage{
makeDiagram(fit, whatLabels = "std", mediationOnly = FALSE,
  residuals = FALSE)
}
\arguments{
\item{fit}{A data.frame. Result of parameterEstimates function of package lavaan}

\item{whatLabels}{What should the edge labels indicate in the path diagram? Choices are c("est","std","name").}

\item{mediationOnly}{Whether or not draw mediation effect only. Default value is FALSE.}

\item{residuals}{Logical, should residuals (and variances) be included in the path diagram?}
}
\description{
Make diagram euation
}
