% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semDiagram.R
\name{makeDiagram}
\alias{makeDiagram}
\title{Make diagram euation}
\usage{
makeDiagram(
  fit,
  labels = NULL,
  graphOptions = list(rankdir = "LR", ranksep = "0.75", nodesep = "0.25", overlap =
    "true", fontsize = "10"),
  nodeOptions = list(),
  edgeOptions = list(color = "black"),
  whatLabels = "std",
  mediationOnly = FALSE,
  regression = TRUE,
  indirect = FALSE,
  secondIndirect = FALSE,
  total = FALSE,
  residuals = FALSE
)
}
\arguments{
\item{fit}{A data.frame. Result of parameterEstimates function of package lavaan}

\item{labels}{An optional named list of variable labels fit object of class lavaan}

\item{graphOptions}{A named list of graph options for Diagrammer syntax}

\item{nodeOptions}{A named list of node options for Diagrammer syntax}

\item{edgeOptions}{A named list of edge options for Diagrammer syntax.}

\item{whatLabels}{What should the edge labels indicate in the path diagram? Choices are c("est","std","name").}

\item{mediationOnly}{Whether or not draw mediation effect only. Default value is FALSE.}

\item{regression}{Whether or not draw regression. Default value is TRUE.}

\item{indirect}{Whether or not draw indirect effects. Default value is FALSE.}

\item{secondIndirect}{Whether or not draw 2nd indirect effects. Default value is FALSE.}

\item{total}{Whether or not draw total effect. Default value is FALSE.}

\item{residuals}{Logical, should residuals (and variances) be included in the path diagram?}
}
\description{
Make diagram euation
}
