% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semDiagram.R
\name{makeSameRank}
\alias{makeSameRank}
\title{make same rank}
\usage{
makeSameRank(equation, temp)
}
\arguments{
\item{equation}{semequation to add}

\item{temp}{character vector}
}
\description{
make same rank
}
