% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsSummary.R
\name{modelsSummary}
\alias{modelsSummary}
\title{Make Model Coef Summary}
\usage{
modelsSummary(fit)
}
\arguments{
\item{fit}{A list of objects of class lm}
}
\value{
A data.frame
}
\description{
Make Model Coef Summary
}
\examples{
fit1=lm(mpg~wt,data=mtcars)
fit2=lm(mpg~wt*hp,data=mtcars)
fit3=lm(mpg~wt*hp*am,data=mtcars)
modelsSummary(list(fit1))
modelsSummary(list(fit1,fit2))
modelsSummary(list(fit1,fit2,fit3))
}
