% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsSummaryTable.R, R/modelsSummaryTable2.R
\name{modelsSummaryTable}
\alias{modelsSummaryTable}
\title{Make Model Coef Table}
\usage{
modelsSummaryTable(x)

modelsSummaryTable(x)
}
\arguments{
\item{x}{object of class modelSummary}

\item{x}{object of class modelSummary}
}
\value{
A flextable

A flextable
}
\description{
Make Model Coef Table

Make Model Coef Table
}
