% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmacroModel.R
\name{pmacroModel}
\alias{pmacroModel}
\title{draw conceptual diagram of process macro model}
\usage{
pmacroModel(no = 1, labels = list())
}
\arguments{
\item{no}{process macro model number}

\item{labels}{A character list}
}
\description{
draw conceptual diagram of process macro model
}
