% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripleInteraction.R
\name{tripleEquation}
\alias{tripleEquation}
\title{Make equation with triple interaction}
\usage{
tripleEquation(X = NULL, M = NULL, Y = NULL, vars, site,
  suffix = 0, covar = NULL, mode = 0)
}
\arguments{
\item{M}{Name of mediator}

\item{Y}{Name of dependent variable}

\item{vars}{variable names to be interact}

\item{suffix}{A number}

\item{covar}{A list of covariates}

\item{mode}{A number}

\item{dep}{name of dependent variable}
}
\description{
Make equation with triple interaction
}
\examples{
X="negemot";M="ideology";Y="govact";vars=c("sex","age");site=c("a","b","c");suffix=0
covar=list(name=c("C1","C2","C3"),label=c("ese","sex","tenure"),site=list(c("M","Y"),"Y","Y"))
cat(tripleEquation(X=X,M=M,Y=Y,vars=vars,site=site))
cat(tripleEquation(X=X,M=M,Y=Y,vars=vars,site=site,covar=covar))
cat(tripleEquation(X=X,M=M,Y=Y,vars=vars,site=site,mode=1))
cat(tripleEquation(X=X,M=M,Y=Y,vars=vars,site=site,covar=covar,mode=1))
}
